+++ src/components/user/work-info/WorkInfo.css
@@ -1,5 +1,3 @@
-@import url('../../../styles/global.css');
-
 /* WorkDetail HEADER */
 .work-header {
   width: calc(var(--space-large) * 42);
