+++ src/pages/user/work-detail/WorkDetail.js
@@ -1,85 +1,11 @@
 import './WorkDetail.css';
 import axios from 'axios';
 import { WorkInfo } from '../../../../src/components/user/work-info/WorkInfo';
+import { generateCalendar } from '../../../utils/generateCalendar';
 
 let currentYear;
 let currentMonth;
 let filteredUsers = []; // 필터링된 사용자 데이터
-
-const generateCalendar = (container, year, month) => {
-  const calendarContainer = container.querySelector('.work-calendar');
-  const calendarTitle = container.querySelector('.calendar-title');
-  calendarContainer.innerHTML = '';
-
-  const daysOfWeek = ['일', '월', '화', '수', '목', '금', '토'];
-  daysOfWeek.forEach(day => {
-    const dayElement = document.createElement('div');
-    dayElement.className = 'calendar-day';
-    dayElement.textContent = day;
-    calendarContainer.appendChild(dayElement);
-  });
-
-  const date = new Date(year, month, 1);
-  const firstDay = date.getDay(); // 첫 번째 날의 요일
-  const totalDays = new Date(year, month + 1, 0).getDate(); // 해당 월의 총 일수
-
-  // 비어 있는 셀 추가
-  for (let i = 0; i < firstDay; i++) {
-    const emptyElement = document.createElement('div');
-    emptyElement.className = 'calendar-day';
-    calendarContainer.appendChild(emptyElement);
-  }
-
-  // 각 날짜에 대한 셀 생성
-  for (let i = 1; i <= totalDays; i++) {
-    const dayElement = document.createElement('div');
-    dayElement.className = 'calendar-day';
-    dayElement.textContent = i;
-
-    const userDataForDate = filteredUsers.filter(user => {
-      const punchDate = new Date(user.punch_date);
-      return (
-        punchDate.getDate() === i &&
-        punchDate.getMonth() === month &&
-        punchDate.getFullYear() === year
-      );
-    });
-
-    // 사용자 데이터가 있을 경우
-    if (userDataForDate.length > 0) {
-      const userData = userDataForDate[0];
-      dayElement.innerHTML += `
-      <div class="cal-punch-group">
-        ${userData.punch_in ? `<p class="cal-punch-in-time">출근</p>` : ''}
-        ${userData.break_out ? `<p class="cal-break-out">외출</p>` : ''}
-      </div>
-    `;
-
-      // 퇴근과 복귀가 있는 경우
-      if (userData.punch_out || userData.break_in) {
-        dayElement.innerHTML += `
-        <div class="cal-punch-group">
-          ${userData.punch_out ? `<p class="cal-punch-out-time">퇴근</p>` : ''}
-          ${userData.break_in ? `<p class="cal-break-in">복귀</p>` : ''}
-        </div>
-      `;
-      }
-    }
-
-    calendarContainer.appendChild(dayElement);
-  }
-
-  // 디자인형식을 맞추기 위함
-  const totalCells = 35;
-  const remainingCells = totalCells - (firstDay + totalDays);
-  for (let i = 0; i < remainingCells; i++) {
-    const emptyElement = document.createElement('div');
-    emptyElement.className = 'calendar-day';
-    calendarContainer.appendChild(emptyElement);
-  }
-
-  calendarTitle.textContent = `${year}년 ${month + 1}월`;
-};
 
 const updatePunchInfo = (container, selectedDate) => {
   const punchInfoTime = container.querySelector('.punch-info-time');
@@ -95,7 +21,7 @@ const updatePunchInfo = (container, selectedDate) => {
     return (
       punchDate.toISOString().split('T')[0] ===
       dayBeforeSelectedDate.toISOString().split('T')[0]
-    ); // ISO 형식으로 비교 형식 :2017-03-16T17:40:00+09:00
+    ); // ISO 형식으로 비교
   });
 
   if (selectedUserData) {
@@ -111,14 +37,14 @@ const updatePunchInfo = (container, selectedDate) => {
     punchInfoTime.innerHTML = `
       <p class="punch-in-time">${punchInTime.getHours()}시 ${punchInTime.getMinutes()}분</p>
       <p class="punch-out-time">${punchOutTime.getHours()}시 ${punchOutTime.getMinutes()}분</p>
-      ${breakOutTime ? `<p class="break-outtime">${breakOutTime.getHours()}시 ${breakOutTime.getMinutes()}분</p>` : ''}
-      ${breakInTime ? `<p class="break-in-time">${breakInTime.getHours()}시 ${breakInTime.getMinutes()}분</p>` : ''}
+      ${breakOutTime ? `<p class="break-outtime">${breakOutTime.getHours()}시 ${breakOutTime.getMinutes()}분</p>` : `<p class="break-outtime">--시 --분</p>`}
+      ${breakInTime ? `<p class="break-in-time">${breakInTime.getHours()}시 ${breakInTime.getMinutes()}분</p>` : `<p class="break-in-time">--시 --분</p>`}
     `;
   } else {
     punchInfoTime.innerHTML = `
       <p class="punch-in-time">--시 --분</p>
       <p class="punch-out-time">--시 --분</p>
-      <p class="break-ot-time">--시 --분</p>
+      <p class="break-outtime">--시 --분</p>
       <p class="break-in-time">--시 --분</p>
     `;
   }
@@ -145,7 +71,7 @@ export const RenderUserWorkDetail = async container => {
 
   // WorkInfo 컴포넌트 호출 및 HTML 삽입
   const workInfoHTML = await WorkInfo(specificUserId);
-  filteredUsers = await fetchFilteredUsers(specificUserId); // 필터링된 사용자 데이터 가져오기
+  filteredUsers = await fetchFilteredUsers(specificUserId);
 
   container.innerHTML = `
     ${workInfoHTML}
@@ -159,7 +85,8 @@ export const RenderUserWorkDetail = async container => {
     </div>
   `;
 
-  generateCalendar(container, currentYear, currentMonth);
+  // generateCalendar 호출 시 filteredUsers 전달
+  generateCalendar(container, currentYear, currentMonth, filteredUsers);
   updatePunchInfo(container, today.toISOString().split('T')[0]); // 오늘 날짜로 초기화
 
   // 아이콘 클릭
@@ -170,7 +97,7 @@ export const RenderUserWorkDetail = async container => {
     } else {
       currentMonth -= 1;
     }
-    generateCalendar(container, currentYear, currentMonth);
+    generateCalendar(container, currentYear, currentMonth, filteredUsers); // 필터링된 사용자 데이터 전달
     updatePunchInfo(container, container.querySelector('.punch-date').value); // 날짜 업데이트
   });
 
@@ -181,13 +108,13 @@ export const RenderUserWorkDetail = async container => {
     } else {
       currentMonth += 1;
     }
-    generateCalendar(container, currentYear, currentMonth);
-    updatePunchInfo(container, container.querySelector('.punch-date').value); // 날짜 업데이트
+    generateCalendar(container, currentYear, currentMonth, filteredUsers);
+    updatePunchInfo(container, container.querySelector('.punch-date').value);
   });
 
   // 날짜 선택
   container.querySelector('.punch-date').addEventListener('change', event => {
     const selectedDate = event.target.value;
-    updatePunchInfo(container, selectedDate); // 해당 날짜의 정보 업데이트
+    updatePunchInfo(container, selectedDate);
   });
 };
